Restore old sns test
